Created Container Extension to load the resource.
